requirements.txt
# Core Dependencies
numpy>=1.26.0
torch>=2.2.0
triton>=2.2.0

# AI Kernel Generation
openai>=1.44.0
deepseek-sdk>=0.2.1

# Hardware Detection
py-cpuinfo>=9.0.0
psutil>=5.9.8
pyopencl>=2024.2

# Performance and Profiling
tqdm>=4.66.4
tabulate>=0.9.0

# Caching and Storage
orjson>=3.10.7
filelock>=3.15.0

# Utilities and Developer Tools
rich>=13.7.1
coloredlogs>=15.0.1
click>=8.1.7

# Type Checking and Packaging
typing-extensions>=4.12.0
packaging>=24.1

# Testing
pytest>=8.3.3
pytest-benchmark>=4.0.0

# Optional Developer Enhancements
ipython>=8.27.0
jupyter>=1.1.1

# Optional: NVIDIA/AMD backend support
cupy-cuda12x>=13.0.0 ; platform_system=="Linux"
rocm-py>=6.0.1
